+++ src/wallet/money/CurrencyUpdaterJSONBuilder.java
@@ -67,7 +67,7 @@ public class CurrencyUpdaterJSONBuilder implements CurrencyUpdaterBuilder{
 
         BigDecimal currencyId = BigDecimal.valueOf((long)currencyObject.get("Cur_ID"));
         BigDecimal currencyScale = BigDecimal.valueOf((long)currencyObject.get("Cur_Scale"));
-        BigDecimal ratio = BigDecimal.valueOf((double)currencyObject.get("Cur_OfficialRate"));
+        double ratio = (double)currencyObject.get("Cur_OfficialRate");
 
         JSONObject localCurrencyObject = new JSONObject();
         localCurrencyObject.put("currencyName",currencyString);
